fix: crash when showing goal stats #77
See issue #77 for details but showing stats for goals crashed the page. The issue was that the goal events don't have a .name, only a .event. And trimUrl started with .length...
This commit fixes the crash but the link for the details is still pointing to .../?filters=((is,goal,(null)))
The trimUrl is also now shielded and logging about these errors.